# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.GenerateAPIKeyRequest:
    properties:
      session_id:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API para integração com WhatsApp usando whatsmeow
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Zemeow WhatsApp API
  version: "1.0"
paths:
  /auth/cache/clear:
    post:
      consumes:
      - application/json
      description: Limpa o cache de autenticação, forçando revalidação de todas as
        API keys
      produces:
      - application/json
      responses:
        "200":
          description: Cache limpo com sucesso
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Acesso negado
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Limpar cache
      tags:
      - auth
  /auth/cache/stats:
    get:
      consumes:
      - application/json
      description: Retorna estatísticas de uso do cache de autenticação
      produces:
      - application/json
      responses:
        "200":
          description: Estatísticas do cache
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Acesso negado
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Obter estatísticas do cache
      tags:
      - auth
  /auth/generate:
    post:
      consumes:
      - application/json
      description: Gera uma nova API key para uma sessão
      parameters:
      - description: Dados para geração da API key
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.GenerateAPIKeyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: API key gerada com sucesso
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Dados inválidos
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Gerar API Key
      tags:
      - auth
  /auth/revoke:
    post:
      consumes:
      - application/json
      description: Revoga uma API key existente, tornando-a inválida
      produces:
      - application/json
      responses:
        "200":
          description: API key revogada com sucesso
          schema:
            additionalProperties: true
            type: object
        "400":
          description: API key não encontrada
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Revogar API Key
      tags:
      - auth
  /auth/validate:
    post:
      consumes:
      - application/json
      description: Valida se uma API key é válida e ativa
      produces:
      - application/json
      responses:
        "200":
          description: API key válida
          schema:
            additionalProperties: true
            type: object
        "401":
          description: API key inválida
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Validar API Key
      tags:
      - auth
...
  /sessions/{sessionId}/media:
    delete:
      consumes:
      - application/json
      description: Remove uma mídia do MinIO e atualiza a mensagem
      parameters:
      - description: ID da sessão
        in: path
        name: sessionId
        required: true
        type: string
      - description: Path da mídia no MinIO
        in: query
        name: path
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Mídia removida com sucesso
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Parâmetros inválidos
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Mídia não encontrada
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Deletar mídia
      tags:
      - media
    get:
      consumes:
      - application/json
      description: Obtém uma mídia armazenada no MinIO pelo path
      parameters:
      - description: ID da sessão
        in: path
        name: sessionId
        required: true
        type: string
      - description: Path da mídia no MinIO
        in: query
        name: path
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Informações da mídia
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Parâmetros inválidos
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Mídia não encontrada
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Obter mídia
      tags:
      - media
  /sessions/{sessionId}/media/download:
    get:
      consumes:
      - application/json
      description: Baixa uma mídia armazenada no MinIO
      parameters:
      - description: ID da sessão
        in: path
        name: sessionId
        required: true
        type: string
      - description: Path da mídia no MinIO
        in: query
        name: path
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Arquivo da mídia
          schema:
            type: file
        "400":
          description: Parâmetros inválidos
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Mídia não encontrada
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Baixar mídia
      tags:
      - media
  /sessions/{sessionId}/media/list:
    get:
      consumes:
      - application/json
      description: Lista todas as mídias de uma sessão com paginação
      parameters:
      - description: ID da sessão
        in: path
        name: sessionId
        required: true
        type: string
      - description: 'Página (padrão: 1)'
        in: query
        name: page
        type: integer
      - description: 'Limite por página (padrão: 50, máximo: 100)'
        in: query
        name: limit
        type: integer
      - description: Filtrar por tipo de mídia (image, video, audio, document)
        in: query
        name: type
        type: string
      - description: Filtrar por direção (incoming, outgoing)
        in: query
        name: direction
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Lista de mídias
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Parâmetros inválidos
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Listar mídias da sessão
      tags:
      - media
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"