# internal/api/handlers/auth.go
package handlers

import (
	"github.com/gofiber/fiber/v2"

	"github.com/felipe/zemeow/internal/logger"
)

type AuthHandler struct {
	logger logger.Logger
}

func NewAuthHandler() *AuthHandler {
	return &AuthHandler{
		logger: logger.GetWithSession("auth_handler"),
	}
}

type GenerateAPIKeyRequest struct {
	SessionID string `json:"session_id"`
}

// @Summary Validar API Key
// @Description Valida se uma API key é válida e ativa
// @Tags auth
// @Accept json
// @Produce json
// @Security ApiKeyAuth
// @Success 200 {object} map[string]interface{} "API key válida"
// @Failure 401 {object} map[string]interface{} "API key inválida"
// @Router /auth/validate [post]
func (h *AuthHandler) ValidateAPIKey(c *fiber.Ctx) error {
	return c.JSON(fiber.Map{
		"valid":   true,
		"message": "API validation endpoint",
	})
}

// @Summary Gerar API Key
// @Description Gera uma nova API key para uma sessão
// @Tags auth
// @Accept json
// @Produce json
// @Security ApiKeyAuth
// @Param request body GenerateAPIKeyRequest true "Dados para geração da API key"
// @Success 200 {object} map[string]interface{} "API key gerada com sucesso"
// @Failure 400 {object} map[string]interface{} "Dados inválidos"
// @Router /auth/generate [post]
func (h *AuthHandler) GenerateAPIKey(c *fiber.Ctx) error {
	return c.JSON(fiber.Map{
		"api_key": "generated-api-key",
		"message": "API key generation endpoint",
	})
}

// @Summary Revogar API Key
// @Description Revoga uma API key existente, tornando-a inválida
// @Tags auth
// @Accept json
// @Produce json
// @Security ApiKeyAuth
// @Success 200 {object} map[string]interface{} "API key revogada com sucesso"
// @Failure 400 {object} map[string]interface{} "API key não encontrada"
// @Router /auth/revoke [post]
func (h *AuthHandler) RevokeAPIKey(c *fiber.Ctx) error {
	return c.JSON(fiber.Map{
		"success": true,
		"message": "API key revocation endpoint",
	})
}

// @Summary Obter estatísticas do cache
// @Description Retorna estatísticas de uso do cache de autenticação
// @Tags auth
// @Accept json
// @Produce json
// @Security ApiKeyAuth
// @Success 200 {object} map[string]interface{} "Estatísticas do cache"
// @Failure 403 {object} map[string]interface{} "Acesso negado"
// @Router /auth/cache/stats [get]
func (h *AuthHandler) GetCacheStats(c *fiber.Ctx) error {
	return c.JSON(fiber.Map{
		"cache_stats": "stats",
		"message":     "Cache stats endpoint",
	})
}

// @Summary Limpar cache
// @Description Limpa o cache de autenticação, forçando revalidação de todas as API keys
// @Tags auth
// @Accept json
// @Produce json
// @Security ApiKeyAuth
// @Success 200 {object} map[string]interface{} "Cache limpo com sucesso"
// @Failure 403 {object} map[string]interface{} "Acesso negado"
// @Router /auth/cache/clear [post]
func (h *AuthHandler) ClearCache(c *fiber.Ctx) error {
	return c.JSON(fiber.Map{
		"success": true,
		"message": "Cache clear endpoint",
	})
}

func (h *AuthHandler) extractAPIKey(c *fiber.Ctx) string {

	authHeader := c.Get("Authorization")
	if authHeader != "" {
		if len(authHeader) > 7 && authHeader[:7] == "Bearer " {
			return authHeader[7:]
		}
		return authHeader
	}

	apiKey := c.Get("X-API-Key")
	if apiKey != "" {
		return apiKey
	}

	return c.Query("api_key")
}

func (h *AuthHandler) sendError(c *fiber.Ctx, message, code string, status int) error {
	errorResp := fiber.Map{
		"error":   code,
		"message": message,
		"status":  status,
	}

	return c.Status(status).JSON(errorResp)
}